Revise pure pursuit node
