+++ catkin_ws/src/pure_pursuit/launch/pure_pursuit.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <launch>
 	<arg name="lookahead" default="0.5"/>
-	<node name="pure_pursuit" pkg="pure_pursuit" type="pure_pursuit.py" output="screen" clear_params="true" required="true">
+	<node name="pure_pursuit" pkg="pure_pursuit" type="pure_pursuit_node.py" output="screen" clear_params="true" required="true">
     	<param name="lookahead" value="$(arg lookahead)"/>
     </node>
 </launch>
