+++ catkin_ws/src/pure_pursuit/setup.py
@@ -1,0 +1,12 @@
+## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+# fetch values from package.xml
+setup_args = generate_distutils_setup(
+    packages=['pure_pursuit'],
+    package_dir={'': 'include'},
+)
+
+setup(**setup_args)
