+++ catkin_ws/src/pure_pursuit/src/pure_pursuit_node.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+
+import rospy
+from Pure_pursuit import Pure_pursuit
+
+"""
+This program utilizes pure pursuit to follow a given trajectory.
+"""
+
+if __name__=="__main__":
+    # Tell ROS that we're making a new node.
+    rospy.init_node("pure_pursuit_node",anonymous=False)
+    pure_pursuit = Pure_pursuit()
+    rospy.spin()
